requirements.txt
Babel==2.14.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
flask-babel==4.0.0
Flask-Table==0.5.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
